# api/openapi.yaml
openapi: 3.0.3
info:
  title: "Auth Service API"
  version: "1.0.0"
  description: |
    Authentication and user/business management service.
    
    Provides endpoints for user registration, login, token refresh,
    profile management, user administration and business resources.
    Protected endpoints require a Bearer JWT access token.
servers:
  - url: "http://localhost:8080"
  - url: "https://api.example.com"
paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      description: >-
        Create a new user account. Rate limited: 5 requests per minute per IP.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        "200":
          description: User registered - returns auth tokens in cookies and success message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/login:
    post:
      summary: Login a user
      description: >-
        Authenticate with email/username and password. Rate limited: 5 requests per minute per IP.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/logout:
    delete:
      summary: Logout current user
      description: Invalidate current access/refresh tokens. Uses cookie-based auth.
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successfully logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/refresh:
    get:
      summary: Refresh access token using refresh token
      description: Returns a new access token when provided a valid refresh token (e.g., cookie or Authorization header). Public endpoint but requires refresh token.
      tags: [Auth]
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          description: 'Bearer {refreshToken} (if using header)'
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/profile:
    get:
      summary: Get current user's profile
      description: Return the profile of the authenticated user.
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update current user's profile
      description: Update fields of the authenticated user's profile.
      tags: [Auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete current user's profile
      description: Permanently delete the authenticated user's account.
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: User deleted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/upload-signature:
    get:
      summary: Get pre-signed upload parameters
      description: |
        Returns Cloudinary upload params (signature, timestamp, api_key, etc.) for direct client-to-Cloudinary uploads.
        Client should POST the image file to the returned upload_url with these params; Cloudinary returns secure_url.
        Then pass that secure_url in PUT /profile as profile_pic.
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Upload signature params for direct Cloudinary upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSignature'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/public-key:
    get:
      summary: Get public key
      description: Returns server's RSA public key in PEM format for token verification.
      tags: [Auth]
      responses:
        "200":
          description: Public key (PEM format)
          content:
            application/x-pem-file:
              schema:
                type: string
                example: |
                  -----BEGIN PUBLIC KEY-----
                  ...
                  -----END PUBLIC KEY-----
        "500":
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required: [id, email, username, createdAt]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        username:
          type: string
          example: "johndoe"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        phone:
          type: string
          example: "+1-202-555-0143"
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          nullable: true
    UserCreate:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        profile_pic:
          type: string
          description: Cloudinary secure_url from direct upload. Use GET /upload-signature first, upload to Cloudinary, then pass the returned secure_url here.
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGci..."
        refreshToken:
          type: string
          example: "dGhpcy1pcy1hLXJlZnJlc2gtdG9rZW4"
        expiresIn:
          type: integer
          description: seconds until expiry
          example: 3600
    LoginRequest:
      type: object
      required: [password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
          format: password
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
    SuccessMessage:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
    UploadSignature:
      type: object
      description: Pre-signed params for direct Cloudinary upload
      properties:
        upload_url:
          type: string
          description: Cloudinary upload endpoint
        api_key:
          type: string
        signature:
          type: string
        timestamp:
          type: string
        folder:
          type: string
        public_id:
          type: string
        cloud_name:
          type: string
...
  responses:
    BadRequest:
      description: Validation failed or bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...